Perceptrons, Error Function, Cross Entropy, Sigmoid, Gradient Descent
Linear regression: Linearly arranged datasets. 
Response variable: The dependent variable whose value we want to explain and predict based on the values of independent variable
For instance:if we want to predict the cost of a house (dependent variable) with respect to the independent variable like size of the house. The datasets can be created with price vs size graph and the value can be predicted considering where the value falls in the regression line.
The data doesn't behave as a straight line. It also accounts for error rate.

Classification:
Based on the given datasets and the constraints, to classify the variable in different class is known as classification.
A line can be created to demonstrate the difference between the classes. Different weights and bias can be used to represent different lines in the cartesian coordinate system.
A prediction can be made based on the score achieved from the equation.

Perceptrons: It takes information from the brain. It receives input processes it and generates output based on the predictive linear model.
It plots the line in the graph and check where the point is based on the linear model. The output is in the form of activation function, here, we use the sigmoid function.
It checks the result of the linear combination. If the result is positive, then it provides a linear combination of 1 and vice versa.
Sigmoid Function: Predicting the probabilites of points and represented in percentage instead of "yes" or "no".

Cross Entropy: How does the computer determines which linear model better classifies the data. The summation of logarithmic probabilities of the points is cross entropy.
When points have high summation value then it is not a good model since it demonstrates a much larger error.

Gradient Descent: To minimize the error to obtain better linear model. Subtracting gradient from linear parameters (W1,W1,Bias) to get the least error.